+++ sensors/tachy/leica/__init__.py
@@ -1,0 +1,1 @@
+import leica_tachy_tps1100