+++ sensors/tachy/leica/leica_tachy_tps1100.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+from leica_tachy import LeicaTachy
+
+
+class TPS1100(LeicaTachy):
+    model = "TPS1100"