+++ sensors/tachy/leica/leica_tachy_ts15.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+from leica_tachy import LeicaTachy
+
+
+class TS15(LeicaTachy):
+    model = "TS15"